DDPB-1196-csv-upload-ajax working version

+++ src/AppBundle/Controller/AdminController.php
@@ -7,6 +7,7 @@ use AppBundle\Exception\DisplayableException;
 use AppBundle\Exception\RestClientException;
 use AppBundle\Form as FormDir;
 use AppBundle\Model\Email;
+use AppBundle\Service\CsvUploader;
 use AppBundle\Service\DataImporter\CsvToArray;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
@@ -27,16 +28,16 @@ class AdminController extends AbstractController
     public function indexAction(Request $request)
     {
         $filters = [
-            'limit' => 100,
-            'offset' => $request->get('offset', 'id'),
-            'role_name' => '',
-            'q' => '',
+            'limit'       => 100,
+            'offset'      => $request->get('offset', 'id'),
+            'role_name'   => '',
+            'q'           => '',
             'odr_enabled' => '',
-            'order_by' => 'id',
-            'sort_order' => 'DESC',
+            'order_by'    => 'id',
+            'sort_order'  => 'DESC',
         ];
 
-        $form = $this->createForm(new FormDir\Admin\SearchType(),null, [ 'method' => 'GET']);
+        $form = $this->createForm(new FormDir\Admin\SearchType(), null, ['method' => 'GET']);
         $form->handleRequest($request);
         if ($form->isValid()) {
             $filters = $form->getData() + $filters;
@@ -45,9 +46,9 @@ class AdminController extends AbstractController
         $users = $this->getRestClient()->get("user/get-all?" . http_build_query($filters), 'User[]');
 
         return [
-            'form'        => $form->createView(),
-            'users'        => $users,
-            'filters'     => $filters,
+            'form'    => $form->createView(),
+            'users'   => $users,
+            'filters' => $filters,
         ];
     }
 
@@ -142,8 +143,8 @@ class AdminController extends AbstractController
                 EntityDir\User::ROLE_PA         => 'Public Authority',
             ],
             'roleNameEmptyValue' => $this->get('translator')->trans('addUserForm.roleName.defaultOption', [], 'admin'),
-            'roleNameSetTo'   => $roleNameSetTo, //can't edit current user's role
-            'odrEnabledType' => $user->getRoleName() == EntityDir\User::ROLE_LAY_DEPUTY ? 'checkbox' : 'hidden',
+            'roleNameSetTo'      => $roleNameSetTo, //can't edit current user's role
+            'odrEnabledType'     => $user->getRoleName() == EntityDir\User::ROLE_LAY_DEPUTY ? 'checkbox' : 'hidden',
         ]), $user);
 
         $clients = $user->getClients();
@@ -250,12 +251,12 @@ class AdminController extends AbstractController
     }
 
     /**
-     * @Route("/upload", name="admin_upload")
+     * @Route("/casrec-upload", name="casrec_upload")
      * @Template
      */
     public function uploadUsersAction(Request $request)
     {
-        $chunkSize = 5000;
+        $chunkSize = 1000;
 
         $form = $this->createForm(new FormDir\UploadCsvType(), null, [
             'method' => 'POST',
@@ -278,37 +279,27 @@ class AdminController extends AbstractController
                     ])
                     ->getData();
 
-
-                // truncate records
-                $this->getRestClient()->delete('casrec/truncate');
-                $request->getSession()->getFlashBag()->add(
-                    'notice', 'Existing casrec data truncated prior to upload'
-                );
-
-
-                $added = 0;
-                $errors = [];
-                foreach (array_chunk($data, $chunkSize) as $chunk) {
-                    $compressedData = base64_encode(gzcompress(json_encode($chunk), 9));
+                // small amount of data -> immediate posting and redirect (needed for behat)
+                if (count($data) < $chunkSize) {
+                    $compressedData = CsvUploader::compressData($data);
+                    $this->getRestClient()->delete('casrec/truncate');
                     $ret = $this->getRestClient()->setTimeout(600)->post('casrec/bulk-add', $compressedData);
-                    $added += $ret['added'];
-                    $errors = array_merge($errors, $ret['errors']);
-                }
-
-                // notifications
-                $request->getSession()->getFlashBag()->add(
-                    'notice',
-                    sprintf('%d record uploaded, %d error(s)', $added, count($errors))
-                );
-                if ($errors) {
                     $request->getSession()->getFlashBag()->add(
                         'notice',
-                        implode('<br/>', $errors)
+                        sprintf('%d record uploaded, %d error(s)', $ret['added'], count($ret['errors']))
                     );
+
+                    return $this->redirect($this->generateUrl('casrec_upload'));
                 }
 
+                // big amount of data => redirect with nOfChunks for ajax upload in chunks
+                $chunks = array_chunk($data, $chunkSize);
+                foreach ($chunks as $k => $chunk) {
+                    $compressedData = CsvUploader::compressData($chunk);
+                    $this->get("snc_redis.default")->set('chunk' . $k, $compressedData);
+                }
 
-                return $this->redirect($this->generateUrl('admin_upload'));
+                return $this->redirect($this->generateUrl('casrec_upload', ['nOfChunks' => count($chunks)]));
             } catch (\Exception $e) {
                 $message = $e->getMessage();
                 if ($e instanceof RestClientException && isset($e->getData()['message'])) {
@@ -319,11 +310,13 @@ class AdminController extends AbstractController
         }
 
         return [
+            'nOfChunks'      => $request->get('nOfChunks'),
             'currentRecords' => $this->getRestClient()->get('casrec/count', 'array'),
             'form'           => $form->createView(),
             'maxUploadSize'  => min([ini_get('upload_max_filesize'), ini_get('post_max_size')]),
         ];
     }
+
 
     /**
      * @Route("/pa-upload", name="admin_pa_upload")
@@ -362,7 +355,7 @@ class AdminController extends AbstractController
                 $added = ['users' => [], 'clients' => [], 'reports' => []];
                 $errors = [];
                 foreach (array_chunk($data, $chunkSize) as $chunk) {
-                    $compressedData = base64_encode(gzcompress(json_encode($chunk), 9));
+                    $compressedData = CsvUploader::compressData($chunk);
                     $ret = $this->getRestClient()->setTimeout(600)->post('pa/bulk-add', $compressedData);
                     $added['users'] = array_merge($added['users'], $ret['added']['users']);
                     $added['clients'] = array_merge($added['clients'], $ret['added']['clients']);
